Arreglando bugs

+++ DTOs/Match/MatchPickBansDto.cs
@@ -3,7 +3,7 @@
     public class MatchPickBansDto
     {
         public bool IsPick { get; set; }
-        public short HeroId { get; set; }
+        public int HeroId { get; set; }
         public int Order { get; set; }
         public bool IsRadiant { get; set; }
         public int PlayerIndex { get; set; }
