+++ Models/MatchPickBans.cs
@@ -7,7 +7,7 @@ namespace StratzAPI.Models
         public long Id { get; set; }
         public long? MatchId { get; set; }
         public bool IsPick { get; set; }
-        public short HeroId { get; set; }
+        public int HeroId { get; set; }
         public int Order { get; set; }
         public bool IsRadiant { get; set; }
         public int PlayerIndex { get; set; }
